Got simple forward blend working in the RenderingEngine. Also fixed a bug with directional Lighting normals being calculated in reverse in the shader.
