+++ GameEngineJava/resource/shaders/forward/ambient.vertex
@@ -1,0 +1,12 @@
+#version 330
+
+layout (location = 0) in vec3 position;
+layout (location = 1) in vec2 texCoord;
+
+out vec2 texCoord0;
+uniform mat4 mvp;
+
+void main() {
+    texCoord0 = texCoord;
+    gl_Position = mvp * vec4( position, 1.0 );
+}
